fix(select): render select list in DOM order rather than in a portal
The select list is no longer rendered inside a portal, but after the select input in DOM order. This
ensures that screenreaders will move directly through the list options when focus is on the input
and the user steps through the following content.

BREAKING CHANGE: The `disablePortal` prop has been removed from all 3 Select variants. It is no
longer needed as the component now always behaves as it did before with this prop added.

fix #5171